+++ README.md
@@ -1,2 +1,6 @@
 # CMSC472FinalProject
 Final Project for CMSC472 - Introduction to Deep Learning with Dr. Abhinav Shrivastava
+
+To run the code:
+* run `conda env create -f env.yml` in the CMSC472FinalProject directory
+* run `conda activate 472` to activate the virtual environment
